Fikset transaksjonsfeil i prosessering
