+++ prosessering/src/test/java/no/nav/familie/prosessering/internal/TaskWorkerTest.kt
@@ -1,131 +1,41 @@
 package no.nav.familie.prosessering.internal
 
-import com.ninjasquad.springmockk.MockkBean
 import io.mockk.every
-import no.nav.familie.log.mdc.MDCConstants
-import no.nav.familie.prosessering.TestAppConfig
+import io.mockk.mockk
+import io.mockk.slot
+import io.mockk.verify
 import no.nav.familie.prosessering.domene.Status
+import no.nav.familie.prosessering.domene.Task
 import no.nav.familie.prosessering.domene.Task.Companion.nyTask
 import no.nav.familie.prosessering.domene.TaskRepository
 import no.nav.familie.prosessering.task.TaskStep1
-import no.nav.familie.prosessering.task.TaskStep2
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.slf4j.MDC
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.autoconfigure.flyway.FlywayAutoConfiguration
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
-import org.springframework.data.domain.PageRequest
-import org.springframework.test.context.ContextConfiguration
-import org.springframework.test.context.junit4.SpringRunner
+import java.util.*
 
-@RunWith(SpringRunner::class)
-@ContextConfiguration(classes = [TestAppConfig::class])
-@DataJpaTest(excludeAutoConfiguration = [FlywayAutoConfiguration::class])
+
 class TaskWorkerTest {
 
-    @MockkBean(relaxUnitFun = true)
-    lateinit var task: TaskStep2
+    private val repository: TaskRepository = mockk(relaxed = true)
 
-    @Autowired
-    private lateinit var repository: TaskRepository
+    private val taskStep1 = TaskStep1(repository)
 
-    @Autowired
-    private lateinit var worker: TaskWorker
-
-    @Test
-    fun `skal hente ut alle tasker uavhengig av status`() {
-        val ubehandletTask = nyTask(TaskStep1.TASK_1, "{'a'='b'}")
-        ubehandletTask.status = Status.UBEHANDLET
-        val feiletTask1 = nyTask(TaskStep2.TASK_2, "{'a'='1'}")
-        feiletTask1.status = Status.FEILET
-        val feiletTask2 = nyTask(TaskStep2.TASK_2, "{'a'='1'}")
-        feiletTask2.status = Status.FEILET
-
-        repository.saveAndFlush(ubehandletTask)
-        repository.saveAndFlush(feiletTask1)
-        repository.saveAndFlush(feiletTask2)
-
-        val alleTasks = repository.finnTasksTilFrontend(Status.values().toList(), PageRequest.of(0, 1000));
-        assertThat(alleTasks.size).isEqualTo(3)
-        assertThat(alleTasks.count { it.status == Status.FEILET }).isEqualTo(2)
-        assertThat(alleTasks.count { it.status == Status.UBEHANDLET }).isEqualTo(1)
-    }
-
-    @Test
-    fun `skal hente ut alle tasker gitt en status`() {
-        val ubehandletTask = nyTask(TaskStep1.TASK_1, "{'a'='b'}")
-        ubehandletTask.status = Status.UBEHANDLET
-        val feiletTask1 = nyTask(TaskStep2.TASK_2, "{'a'='1'}")
-        feiletTask1.status = Status.FEILET
-        val feiletTask2 = nyTask(TaskStep2.TASK_2, "{'a'='1'}")
-        feiletTask2.status = Status.FEILET
-
-        repository.saveAndFlush(ubehandletTask)
-        repository.saveAndFlush(feiletTask1)
-        repository.saveAndFlush(feiletTask2)
-
-        val alleTasks = repository.finnTasksTilFrontend(listOf(Status.FEILET), PageRequest.of(0, 1000));
-        assertThat(alleTasks.size).isEqualTo(2)
-        assertThat(alleTasks.count { it.status == Status.FEILET }).isEqualTo(2)
-    }
-
-    @Test
-    fun `skal hente task til frontend`() {
-        MDC.put(MDCConstants.MDC_CALL_ID, "test")
-        val feiletTask1 = nyTask(TaskStep2.TASK_2, "{'a'='1'}")
-        feiletTask1.status = Status.FEILET
-        repository.saveAndFlush(feiletTask1)
-
-        val tasks = repository.finnTasksTilFrontend(listOf(Status.FEILET))
-
-        assertThat(tasks).hasSize(1)
-
-        val task = tasks.first()
-        val id = task.id
-        val tasklogg = repository.finnTaskloggTilFrontend(id)
-
-        assertThat(task.callId).isEqualTo("test")
-        assertThat(tasklogg).hasSize(1)
-        assertThat(task.sistKjørt).isEqualTo(tasklogg.first().opprettetTidspunkt)
-
-        MDC.remove(MDCConstants.MDC_CALL_ID)
-    }
+    private val worker: TaskWorker = TaskWorker(repository, listOf(taskStep1))
 
     @Test
     fun `skal behandle task`() {
-        var task1 =
-                nyTask(TaskStep1.TASK_1, "{'a'='b'}")
-        repository.saveAndFlush(task1)
-        assertThat(task1.status)
-                .isEqualTo(Status.UBEHANDLET)
+        val task1 =
+                nyTask(TaskStep1.TASK_1, "{'a'='b'}").copy(id = 1)
+        every { repository.findById(1) } returns Optional.of(task1)
+        val copy = task1.copy(status = Status.BEHANDLER)
+        every { repository.saveAndFlush(copy) } returns copy
+        assertThat(task1.status).isEqualTo(Status.UBEHANDLET)
 
         worker.doActualWork(task1.id!!)
 
-        task1 = repository.findById(task1.id!!).orElseThrow()
+        verify { repository.saveAndFlush(copy) }
         assertThat(task1.status).isEqualTo(Status.FERDIG)
         assertThat(task1.logg).hasSize(3)
     }
 
-    @Test
-    fun `skal håndtere feil`() {
-        var task2 = nyTask(TaskStep2.TASK_2, "{'a'='b'}")
-        repository.saveAndFlush(task2)
-        assertThat(task2.status).isEqualTo(Status.UBEHANDLET)
-        every { task.doTask(any()) } throws (IllegalStateException())
-
-        worker.doActualWork(task2.id!!)
-        task2 = repository.findById(task2.id!!).orElseThrow()
-        assertThat(task2.status).isEqualTo(Status.KLAR_TIL_PLUKK)
-        assertThat(task2.logg).hasSize(3)
-
-        worker.doActualWork(task2.id!!)
-        task2 = repository.findById(task2.id!!).orElseThrow()
-        assertThat(task2.status).isEqualTo(Status.KLAR_TIL_PLUKK)
-
-        worker.doActualWork(task2.id!!)
-        task2 = repository.findById(task2.id!!).orElseThrow()
-        assertThat(task2.status).isEqualTo(Status.FEILET)
-    }
 }
